Updating text to something less opinionated
I know many people that care a great deal about where materials for your products are sourced or which persons fabricated this product. But I still really like this metaphor, so I propose a change of this example to something that doesn't trigger as many emotions.
This proposal took me just a few minutes to type up, maybe you agree and help me to come up with a good example.