Kindly add this to tic tac toe assignment
I know the code is looking copied and I admit it. But I have learned a lot while copying this code and got, at least, some insight in the OO programming.
This is the first time that I have solved such a problem though my solution was not worth posting.