Update sessions_cookies_authentication.md
Changed the description of Authentication, it erroneously was saying that "The whole point of authentication is to make sure that whoever is requesting to run an action is actually ALLOWED to do so." That in fact describes authorization. It should read that authentication makes sure you are who you say you are. big defference