sublime text 2 -> sublime text
Sublime text 3 is just referred to as Sublime Text now. Since what is on the ST homepage is actually ST3, i'd suggest removing the mention of ST2 to avoid confusion. If you'd rather people use ST2, since 3 is technically still in beta, change the link to https://www.sublimetext.com/2