Resolve Broken Link #5049
I added the links as discussed. I also reached out to Albert Park to inquire whether he would be finishing with part III. He replied that he was planning on it, but since now he knows there is interest, he will move it higher on the priority list. Once he publishes it, I'll update this page. In the meantime I made sure to acknowledge that the two parts only deal with views and controllers.