Adds Josh Archer's solution and fixes malformed markdown links
In addition to Josh Archer's solution, and link to live project, a link to a walkthrough of the code and how the project was written has been included. This should provide beneficial insight to any future students as it contains explanations/advice regarding the project that cannot simply be included in inline comments.

This PR also fixes numerous malformed links affecting the student solutions section of the page.