Remove specific ruby versioning from instructions
The current installation instructions give the specific ruby version installed. Somebody in the forums has sought feedback because their versioning doesn't match that so it would be better to be less specific and only provide a guide to what feedback they should see on the screen. Any version specific information has been replaced.