Updates text in assignments area: terminal & AWS
The lesson was inviting people to skip ahead for installations. Updated text to reflect the fact that installations happen before this lesson.

Removed reference to AWS. Since learners should already have linux installed (or using MacOS), AWS shouldn't be necessary.

Added a step that shows learners how to open a terminal and try a basic command. This isn't explicitly mentioned at the start of the Conquering the Command Line book.