Updates community lesson
I moved the links to the chat and the forum so that learners are reading through our suggestions. Added a link to instagram. Alsp, created a section for suggestions before asking for help and another when actually asking for help. Also adds suggestion about using google or looking at relevant lessons before turning to the community. Also adds the don't ask to ask link.