Fix codeblocks rendering as plaintext on website
Inserted newlines after paras so that code blocks are rendered not as normal text on the main website. Currently, many code blocks in this article are being rendered as plain text on the website because of the lack of newline between a paragraph and commencing code block codes in the markdown file.