Removed repeated paragraph
The paragraph on line 76 that begins with, "Once you start to feel comfortable..." was also verbatim in the motivation section of the lesson. I removed the paragraph from the motivation section since I felt that the paragraph better led into the conclusion, which meant it should be in the Why Odin? section.